Reduce packed TransitionMeta size to 109 bytes
Updated the TransitionMeta struct and related packing/unpacking logic in LibCodec to use 109 bytes instead of 121/122 by reducing provabilityBond and livenessBond fields to 6 bytes each (uint48). Adjusted tests and documentation accordingly. Also changed bondToProver type in LibVerify to uint256 for consistency.

+++ packages/protocol/contracts/layer1/based2/libs/LibVerify.sol
@@ -70,7 +70,7 @@ library LibVerify {
                     break;
                 }
 
-                uint96 bondToProver =
+                uint256 bondToProver =
                     _calcBondToProver(_conf, _trans[i], isFirstTransition) * ONE_GWEI;
                 _rw.creditBond(_trans[i].prover, bondToProver);
